Move Selects HOCs in the hoc folder
