+++ packages/react-vapor/src/components/select/hoc/SelectWithPredicate.tsx
@@ -3,14 +3,14 @@ import {connect} from 'react-redux';
 import {keys} from 'ts-transformer-keys';
 import * as _ from 'underscore';
 
-import {WithServerSideProcessingProps} from '../../hoc/withServerSideProcessing/withServerSideProcessing';
-import {IReactVaporState} from '../../ReactVapor';
-import {callIfDefined} from '../../utils/FalsyValuesUtils';
-import {FlatSelectConnected} from '../flatSelect/FlatSelectConnected';
-import {IFlatSelectOptionProps} from '../flatSelect/FlatSelectOption';
-import {FlatSelectSelectors} from '../flatSelect/FlatSelectSelectors';
-import {IItemBoxProps} from '../itemBox/ItemBox';
-import {ISelectProps} from './SelectConnected';
+import {WithServerSideProcessingProps} from '../../../hoc/withServerSideProcessing/withServerSideProcessing';
+import {IReactVaporState} from '../../../ReactVapor';
+import {callIfDefined} from '../../../utils/FalsyValuesUtils';
+import {FlatSelectConnected} from '../../flatSelect/FlatSelectConnected';
+import {IFlatSelectOptionProps} from '../../flatSelect/FlatSelectOption';
+import {FlatSelectSelectors} from '../../flatSelect/FlatSelectSelectors';
+import {IItemBoxProps} from '../../itemBox/ItemBox';
+import {ISelectProps} from '../SelectConnected';
 
 export interface ISelectWithPredicateOwnProps {
     options: IFlatSelectOptionProps[];
