+++ packages/react-vapor/src/components/select/tests/hoc/MultiSelectWithFilter.spec.tsx
@@ -2,15 +2,16 @@ import {mount, ReactWrapper} from 'enzyme';
 import * as React from 'react';
 import {Provider} from 'react-redux';
 import {Store} from 'redux';
-import {IReactVaporState} from '../../../ReactVapor';
-import {clearState} from '../../../utils/ReduxUtils';
-import {TestUtils} from '../../../utils/tests/TestUtils';
-import {filterThrough} from '../../filterBox/FilterBoxActions';
-import {ItemBox} from '../../itemBox/ItemBox';
-import {toggleSelect} from '../SelectActions';
-import {MultiSelectWithFilter} from '../SelectComponents';
-import {SelectConnected} from '../SelectConnected';
-import {ISelectWithFilterProps} from '../SelectWithFilter';
+
+import {IReactVaporState} from '../../../../ReactVapor';
+import {clearState} from '../../../../utils/ReduxUtils';
+import {TestUtils} from '../../../../utils/tests/TestUtils';
+import {filterThrough} from '../../../filterBox/FilterBoxActions';
+import {ItemBox} from '../../../itemBox/ItemBox';
+import {ISelectWithFilterProps} from '../../hoc/SelectWithFilter';
+import {toggleSelect} from '../../SelectActions';
+import {MultiSelectWithFilter} from '../../SelectComponents';
+import {SelectConnected} from '../../SelectConnected';
 
 describe('Select', () => {
     describe('MultiSelectWithFilter', () => {
