+++ packages/react-vapor/src/components/select/tests/hoc/SelectWithFilterReducers.spec.ts
@@ -1,14 +1,14 @@
-import {StringListActions} from '../../../reusableState/customList/StringListActions';
-import * as StringListReducers from '../../../reusableState/customList/StringListReducers';
-import {IStringListCompositeState} from '../../../reusableState/customList/StringListReducers';
-import {stringListInitialState} from '../../../reusableState/customList/StringListReducers';
-import {IReduxAction} from '../../../utils/ReduxUtils';
-import {clearListBoxOption, selectListBoxOption, unselectListBoxOption} from '../../listBox/ListBoxActions';
+import {StringListActions} from '../../../../reusableState/customList/StringListActions';
+import * as StringListReducers from '../../../../reusableState/customList/StringListReducers';
+import {IStringListCompositeState} from '../../../../reusableState/customList/StringListReducers';
+import {stringListInitialState} from '../../../../reusableState/customList/StringListReducers';
+import {IReduxAction} from '../../../../utils/ReduxUtils';
+import {clearListBoxOption, selectListBoxOption, unselectListBoxOption} from '../../../listBox/ListBoxActions';
 import {
     ISelectWithFilterCompositeState,
     ISelectWithFilterPayload,
     selectWithFilterCompositeReducer,
-} from '../SelectWithFilterReducers';
+} from '../../hoc/SelectWithFilterReducers';
 
 describe('Select', () => {
     describe('SelectWithFilter reducers', () => {
