+++ packages/react-vapor/src/components/select/tests/hoc/SelectWithInfiniteScroll.spec.tsx
@@ -3,11 +3,11 @@ import {shallowWithStore} from 'enzyme-redux';
 import * as React from 'react';
 import * as InfiniteScroll from 'react-infinite-scroll-component';
 
-import {getStoreMock} from '../../../utils/tests/TestUtils';
-import {IItemBoxProps} from '../../itemBox/ItemBox';
-import {selectWithInfiniteScroll, SelectWithInfiniteScrollProps} from '../hoc/SelectWithInfiniteScroll';
-import {ISelectProps} from '../SelectConnected';
-import {ISingleSelectOwnProps, SingleSelectConnected} from '../SingleSelectConnected';
+import {getStoreMock} from '../../../../utils/tests/TestUtils';
+import {IItemBoxProps} from '../../../itemBox/ItemBox';
+import {selectWithInfiniteScroll, SelectWithInfiniteScrollProps} from '../../hoc/SelectWithInfiniteScroll';
+import {ISelectProps} from '../../SelectConnected';
+import {ISingleSelectOwnProps, SingleSelectConnected} from '../../SingleSelectConnected';
 
 describe('SelectWithInfiniteScroll', () => {
     const SingleSelectWithInfiniteScroll = selectWithInfiniteScroll(SingleSelectConnected);
